Preparação do deploy feita na mentoria

+++ pages/course/[id].tsx
@@ -15,6 +15,15 @@ const CoursePage = function () {
   const [course, setCourse] = useState<CourseType>();
   const router = useRouter();
   const { id } = router.query;
+  const [loading, setLoading] = useState(true);
+
+  useEffect(() => {
+    if (!sessionStorage.getItem('onebitflix-token')) {
+      router.push('/login');
+    } else {
+      setLoading(false);
+    }
+  }, []);
 
   const getCourse = async function () {
     if (typeof id !== 'string') return;
@@ -33,16 +42,18 @@ const CoursePage = function () {
   }, [id]);
 
   const handleLikeCourse = async () => {
+    if (typeof id !== 'string') return;
     if (liked === true) {
       await courseService.removeLike(id);
       setLiked(false);
     } else {
-      await courseService.like(id)
+      await courseService.like(id);
       setLiked(true);
     }
   };
 
   const handleFavCourse = async () => {
+    if (typeof id !== 'string') return;
     if (favorited === true) {
       await courseService.removeFav(id);
       setFavorited(false);
@@ -52,7 +63,9 @@ const CoursePage = function () {
     }
   };
 
-  if (course === undefined) return <PageSpinner />;
+  if (loading) {
+    return <PageSpinner />;
+  }
 
   return (
     <>
